datazen/requirements.txt
Jinja2
pyyaml
cerberus
